+++ app/src/main/java/org/fitchfamily/android/dejavu/BackendService.kt
@@ -554,6 +554,7 @@ class BackendService : LocationBackendService() {
      * Call back method entered when Android has completed a scan for WiFi emitters in
      * the area.
      */
+    @SuppressLint("MissingPermission")
     @Synchronized
     private fun onWiFisChanged() {
         val scanResults = wifiManager.scanResults